+++ src/main/java/kitchenpos/domain/Order.java
@@ -15,6 +15,7 @@ import javax.persistence.Id;
 import javax.persistence.JoinColumn;
 import javax.persistence.ManyToOne;
 import javax.persistence.Table;
+import kitchenpos.domain.table.OrderTable;
 import kitchenpos.support.domain.BaseEntity;
 
 @Table(name = "orders")
