+++ src/main/java/kitchenpos/domain/TableGroup.java
@@ -10,6 +10,7 @@ import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
 import javax.persistence.Id;
 import javax.persistence.OneToMany;
+import kitchenpos.domain.table.OrderTable;
 import kitchenpos.support.domain.BaseEntity;
 import org.springframework.util.CollectionUtils;
