+++ src/main/java/kitchenpos/domain/table/OrderTableChangedEvent.java
@@ -1,4 +1,4 @@
-package kitchenpos.domain;
+package kitchenpos.domain.table;
 
 public class OrderTableChangedEvent {
