+++ src/main/java/kitchenpos/domain/table/OrderTableRepository.java
@@ -1,7 +1,6 @@
-package kitchenpos.dao;
+package kitchenpos.domain.table;
 
 import java.util.List;
-import kitchenpos.domain.OrderTable;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
