+++ src/main/java/kitchenpos/dto/OrderCreateRequest.java
@@ -6,7 +6,7 @@ import java.util.List;
 import java.util.stream.Collectors;
 import kitchenpos.domain.Order;
 import kitchenpos.domain.OrderLineItem;
-import kitchenpos.domain.OrderTable;
+import kitchenpos.domain.table.OrderTable;
 
 public class OrderCreateRequest {
     private Long orderTableId;
