+++ src/test/java/kitchenpos/domain/OrderTest.java
@@ -7,6 +7,7 @@ import static kitchenpos.fixture.OrderLineItemFixture.주문_항목;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 import java.util.List;
+import kitchenpos.domain.table.OrderTable;
 import org.junit.jupiter.api.DisplayNameGeneration;
 import org.junit.jupiter.api.DisplayNameGenerator;
 import org.junit.jupiter.api.Test;
