+++ src/test/java/kitchenpos/domain/table/OrderTableTest.java
@@ -1,8 +1,10 @@
-package kitchenpos.domain;
+package kitchenpos.domain.table;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
+import java.time.LocalDateTime;
+import kitchenpos.domain.TableGroup;
 import org.junit.jupiter.api.DisplayNameGeneration;
 import org.junit.jupiter.api.DisplayNameGenerator;
 import org.junit.jupiter.api.Test;
@@ -36,7 +38,7 @@ class OrderTableTest {
     @Test
     void 테이블의_상태를_변경하는_경우_이미_단체지정이_되어있는_테이블이라면_예외를_던진다() {
         // given
-        OrderTable orderTable = new OrderTable(new TableGroup(), 0, true);
+        OrderTable orderTable = new OrderTable(new TableGroup(LocalDateTime.now()), 0, true);
 
         // expect
         assertThatThrownBy(() -> orderTable.changeEmpty(true))
@@ -47,7 +49,7 @@ class OrderTableTest {
     @Test
     void 단체지정을_해제한다() {
         // given
-        OrderTable orderTable = new OrderTable(new TableGroup(), 0, true);
+        OrderTable orderTable = new OrderTable(new TableGroup(LocalDateTime.now()), 0, true);
 
         // when
         orderTable.clearTableGroup();
@@ -59,10 +61,10 @@ class OrderTableTest {
     @Test
     void 단체지정을_할_때_이미_단체지정이_되어있으면_예외를_던진다() {
         // given
-        OrderTable orderTable = new OrderTable(new TableGroup(), 0, true);
+        OrderTable orderTable = new OrderTable(new TableGroup(LocalDateTime.now()), 0, true);
 
         // expect
-        assertThatThrownBy(() -> orderTable.changeTableGroup(new TableGroup()))
+        assertThatThrownBy(() -> orderTable.changeTableGroup(new TableGroup(LocalDateTime.now())))
                 .isInstanceOf(IllegalArgumentException.class)
                 .hasMessage("비어있지 않거나, 이미 단체 지정이 된 테이블은 단체 지정을 할 수 없습니다.");
     }
@@ -73,7 +75,7 @@ class OrderTableTest {
         OrderTable orderTable = new OrderTable(null, 0, false);
 
         // expect
-        assertThatThrownBy(() -> orderTable.changeTableGroup(new TableGroup()))
+        assertThatThrownBy(() -> orderTable.changeTableGroup(new TableGroup(LocalDateTime.now())))
                 .isInstanceOf(IllegalArgumentException.class)
                 .hasMessage("비어있지 않거나, 이미 단체 지정이 된 테이블은 단체 지정을 할 수 없습니다.");
     }
@@ -84,7 +86,7 @@ class OrderTableTest {
         OrderTable orderTable = new OrderTable(null, 0, true);
 
         // when
-        orderTable.changeTableGroup(new TableGroup());
+        orderTable.changeTableGroup(new TableGroup(LocalDateTime.now()));
 
         // then
         assertThat(orderTable.isEmpty()).isFalse();
