+++ src/test/java/kitchenpos/fixture/OrderTableFixture.java
@@ -1,7 +1,7 @@
 package kitchenpos.fixture;
 
-import kitchenpos.domain.OrderTable;
 import kitchenpos.domain.TableGroup;
+import kitchenpos.domain.table.OrderTable;
 
 public class OrderTableFixture {
