+++ src/test/java/kitchenpos/fixture/TableGroupFixture.java
@@ -6,8 +6,8 @@ import static java.util.stream.Collectors.toList;
 import java.time.LocalDateTime;
 import java.util.Arrays;
 import java.util.List;
-import kitchenpos.domain.OrderTable;
 import kitchenpos.domain.TableGroup;
+import kitchenpos.domain.table.OrderTable;
 import kitchenpos.dto.TableGroupRequest;
 import kitchenpos.dto.TableGroupRequest.OrderTableIdRequest;
